backend/requirements.txt
# ================================================================
# DÉPENDANCES BACKEND - BaraCorrespondance AI
# ================================================================

# Flask Core
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-Cors==4.0.0
Flask-Mail==0.9.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5

# Authentication
Flask-JWT-Extended==4.6.0
bcrypt==4.1.1

# Validation
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
email-validator==2.1.0

# File Processing - CV Parsing
python-docx==1.1.0
pdfplumber==0.10.3
PyPDF2==3.0.1

# NLP & AI pour analyse CV
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2

# Google Gemini (Gratuit et performant)
google-generativeai==0.3.2

# OpenAI (Legacy)
openai==1.6.1

# Image Generation (Affiches)
Pillow==10.1.0
reportlab==4.0.7

# Utils
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1

# Testing
pytest==7.4.3
pytest-flask==1.3.0
